+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/dto/request/CreateOutingTypeRequest.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.domain.outing.dto
+package team.aliens.dms.domain.outing.dto.request
 
 data class CreateOutingTypeRequest(
     val title: String
