+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/dto/response/ApplyOutingResponse.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.domain.outing.dto
+package team.aliens.dms.domain.outing.dto.response
 
 import java.util.UUID
