+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/usecase/CreateOutingTypeUseCase.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.domain.outing.usecase
 
 import team.aliens.dms.common.annotation.UseCase
 import team.aliens.dms.common.service.security.SecurityService
-import team.aliens.dms.domain.outing.dto.CreateOutingTypeRequest
+import team.aliens.dms.domain.outing.dto.request.CreateOutingTypeRequest
 import team.aliens.dms.domain.outing.model.OutingType
 import team.aliens.dms.domain.outing.service.OutingService
