+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/usecase/GetAllOutingTypeTitlesUseCase.kt
@@ -2,7 +2,7 @@ package team.aliens.dms.domain.outing.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
 import team.aliens.dms.common.service.security.SecurityService
-import team.aliens.dms.domain.outing.dto.GetAllOutingTypeTitlesResponse
+import team.aliens.dms.domain.outing.dto.response.GetAllOutingTypeTitlesResponse
 import team.aliens.dms.domain.outing.service.GetOutingService
 
 @ReadOnlyUseCase
