+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/usecase/GetOutingAvailableTimesUseCase.kt
@@ -1,7 +1,7 @@
 package team.aliens.dms.domain.outing.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
-import team.aliens.dms.domain.outing.dto.OutingAvailableTimesResponse
+import team.aliens.dms.domain.outing.dto.response.OutingAvailableTimesResponse
 import team.aliens.dms.domain.outing.service.OutingService
 import java.time.DayOfWeek
