+++ dms-core/src/main/kotlin/team/aliens/dms/domain/outing/usecase/GetOutingHistoryDetailsUseCase.kt
@@ -1,7 +1,7 @@
 package team.aliens.dms.domain.outing.usecase
 
 import team.aliens.dms.common.annotation.ReadOnlyUseCase
-import team.aliens.dms.domain.outing.dto.OutingHistoryDetailsResponse
+import team.aliens.dms.domain.outing.dto.response.OutingHistoryDetailsResponse
 import team.aliens.dms.domain.outing.service.OutingService
 import team.aliens.dms.domain.student.service.StudentService
 import java.util.UUID
